Refactor settings to a different folder
