+++ packages/unmock-core/src/settings/index.ts
@@ -1,0 +1,2 @@
+export { AllowedHosts } from "./allowedHosts";
+export { BooleanSetting } from "./boolean";
