+++ packages/unmock-core/src/settings/utils.ts
@@ -1,0 +1,17 @@
+import { escapeRegExp } from "lodash";
+
+export const whitelistToRegex = (
+  whitelist?: Array<string | RegExp>,
+): RegExp[] =>
+  whitelist === undefined
+    ? []
+    : whitelist.map((item: any) =>
+        item instanceof RegExp
+          ? item
+          : new RegExp(
+              `^${item
+                .split(/\*+/)
+                .map(escapeRegExp)
+                .join(".*")}$`,
+            ),
+      );
